third_party/WebKit/LayoutTests/http/tests/inspector-protocol/network/disable-interception-midway-expected.txt

Tests interception blocking, modification of network fetches.
Test started
Network agent enabled
Request interception enabled
Page agent enabled
Runtime agent enabled
Network.requestIntercepted ID 1 GET resource-iframe.html type: Document
allowRequest ID 1
Network.responseReceived resource-iframe.html 200 text/html
Network.requestIntercepted ID 2 GET i-dont-exist.css type: Stylesheet
Network.responseReceived i-dont-exist.css 404 text/html
Network.loadingFailed i-dont-exist.css 
Network.requestIntercepted ID 3 GET script.js type: Script
Network.responseReceived script.js 404 text/html
Network.loadingFailed script.js 
Network.requestIntercepted ID 4 GET script2.js type: Script
----- disableRequestInterception -----
Network.responseReceived script2.js 404 text/html
Network.loadingFailed script2.js 
Page.frameStoppedLoading
Network.responseReceived post-echo.pl 200 text/plain
Post - ECHO SUCCESS!
